\name{modelpsm}
\alias{modelpsm}
\title{modelpsm
}
\description{
The JAGS model to feed in BPSM. It is an internal function called by BPSM.
}
\usage{modelpsm()}
\author{
Weihua An, Huizi Xu, and Zhida Zheng, Indiana University Bloomington.
}

\seealso{
\code{\link{bpsm}, \link{bpsr}, \link{modelpsr}, \link{Match}, \link{sortps}}
}
