\name{modelpsr}
\alias{modelpsr}
\title{modelpsr}
\description{
The JAGS model to feed in BPSR. It is an internal function called by BPSR.
}
\usage{modelpsr()}
\author{
Weihua An, Huizi Xu, and Zhida Zheng, Indiana University Bloomington.
}

\seealso{
\code{\link{bpsm}, \link{bpsr}, \link{modelpsm}, \link{Match}, \link{sortps}}
}